Squashed in curve-curve-intersection-cubic-quad-line:
    * Added tests
    * Enabled intersections between lines and other curve types
    * Added tests for intersections between quadratic bezier curves
    * Moved ParamCurveBezierClipping into new file
