+++ src/lib.rs
@@ -90,6 +90,7 @@ mod insets;
 mod line;
 mod mindist;
 mod param_curve;
+mod param_curve_clipping;
 mod point;
 mod quadbez;
 mod quadspline;
@@ -113,6 +114,7 @@ pub use crate::ellipse::*;
 pub use crate::insets::*;
 pub use crate::line::*;
 pub use crate::param_curve::*;
+pub use crate::param_curve_clipping::*;
 pub use crate::point::*;
 pub use crate::quadbez::*;
 pub use crate::quadspline::*;
